+++ tests/test_utils.py
@@ -1,9 +1,11 @@
 import datetime as dt
 
 import pytest
+from aiopg.sa import SAConnection
 from pydantic import BaseModel
 
-from repka.utils import model_to_primitive
+from repka.utils import model_to_primitive, create_async_db_connection
+from tests.conftest import DB_URL
 
 
 class MyModel(BaseModel):
@@ -34,3 +36,10 @@ def test_model_to_primitive_excludes_id(model: MyModel) -> None:
 def test_model_to_primitive_excludes_fields_from_list(model: MyModel) -> None:
     dict_ = model_to_primitive(model, exclude=["title", "created"])
     assert "title" not in dict_ and "created" not in dict_
+
+
+@pytest.mark.asyncio
+async def test_create_async_db_connection() -> None:
+    async with create_async_db_connection(DB_URL) as connection:
+        conn: SAConnection = connection
+        assert conn.connection.status